Fixed the collision checker in GlobalPlanner so that it will work correctly with MultiTraitOccupancyGrid and AStarCarrot #105

+++ aero_path_planning/src/planners/GlobalPlanner.cpp
@@ -441,7 +441,7 @@ bool GlobalPlanner::checkCollision(const tf::Point& point, const occupancy_grid:
 	bool collision = false;
 	try
 	{
-		if(map.getPointTrait(point.x(), point.y())==occupancy_grid::utilities::CellTrait::OBSTACLE)
+		if(map.getPointTrait((unsigned int)point.x(), (unsigned int)point.y())==occupancy_grid::utilities::CellTrait::OBSTACLE)
 		{
 			collision = true;
 		}
